requirements.txt
streamlit
langchain
langchain-community
langchain-groq
beautifulsoup4
lxml
python-dotenv